Fix: Move component definition out of the parent component
